% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unleash.R
\name{unleash}
\alias{unleash}
\title{UNimodally Leveraging Empirical-null with Adaptive SHrinkage}
\usage{
unleash(betahat, sebetahat, mixcompdist = c("uniform", "halfuniform",
  "normal", "+uniform", "-uniform"), df = NULL, pen = 1, maxiter = 500,
  tol = 10^-6, plot_update = FALSE, ...)
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{sebetahat}{a p vector of corresponding standard errors}

\item{mixcompdist}{distribution of components in mixture
("uniform","halfuniform" or "normal"; "+uniform" or
"-uniform"), the default is "uniform". If you believe your
effects may be asymmetric, use "halfuniform". If you want to
allow only positive/negative effects use "+uniform"/"-uniform".
The use of "normal" is permitted only if df=NULL.}

\item{df}{appropriate degrees of freedom for (t) distribution of
betahat/sebetahat, default is NULL which is actually treated as
infinity (Gaussian)}

\item{pen}{The variance inflation penalty.}

\item{maxiter}{The maximum number of iterations to run.}

\item{tol}{The stopping criterion.}

\item{plot_update}{A logical. Should we plot the updates?}

\item{...}{Further arguments to be passed to
\code{\link{ash.workhorse}}.}
}
\description{
This is a wrapper for \code{\link{ash}} with a variance inflation term that
expands the variances to as large a value as possible while allowing
the prior to be unimodal.
}
\author{
David Gerard
}
