% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unleash.R
\name{var_inflate_obj}
\alias{var_inflate_obj}
\title{Wrapper for \code{\link{calc_loglik}}}
\usage{
var_inflate_obj(xi, fitted_g, data, pen = 1)
}
\arguments{
\item{xi}{The variance inflation parameter.}

\item{fitted_g}{The fitted g}

\item{data}{a data object, see set_data}

\item{pen}{The variance inflation penalty.}
}
\description{
Wrapper for \code{\link{calc_loglik}}
}
\author{
David Gerard
}
